requirements.txt
alembic==1.4.3
and==0.1.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap==0.2.4
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-summernote==0.8.11.6
django-tinymce==3.3.0
dominate==2.6.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown2==2.3.10
MarkupSafe==1.1.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pictures==0.7
Pillow==8.1.0
prompt-toolkit==3.0.10
protobuf==3.6.1
psycopg2==2.8.6
ptyprocess==0.7.0
pycairo==1.16.2
pycups==1.9.73
Pygments==2.7.3
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyperclip==1.8.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-editor==1.0.4
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
systemd-python==234
traitlets==5.0.5
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.2
visitor==0.1.3
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
WTForms==2.3.3
xkit==0.0.0
yapf==0.30.0
